training_service/requirements.txt
fastapi
uvicorn[standard]
sse-starlette
python-multipart
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
Pillow
py-cpuinfo
